+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Parse `HTMLElement#innerHTML` assignments to actual elements on-the-fly on the server
 * Allow plain HTML custom Element templates to be rendered synchronously upon creation
 * During a render, element options are only applied when closing the element, or the current element is accessed using `$0`
+* Fix `Hawkejs.getFirstElement(entries)` returning text nodes
 
 ## 2.0.4 (2020-10-22)